address/model: change name from AddressBook to PersonnelDatabase
