+++ src/main/java/seedu/address/model/PersonnelDatabase.java
@@ -11,10 +11,10 @@ import seedu.address.model.person.Person;
 import seedu.address.model.person.UniquePersonList;
 
 /**
- * Wraps all data at the address-book level
+ * Wraps all data at the personnel database level
  * Duplicates are not allowed (by .isSamePerson comparison)
  */
-public class AddressBook implements ReadOnlyAddressBook {
+public class PersonnelDatabase implements ReadOnlyPersonnelDatabase {
 
     private final UniquePersonList persons;
     private final InvalidationListenerManager invalidationListenerManager = new InvalidationListenerManager();
@@ -30,12 +30,12 @@ public class AddressBook implements ReadOnlyAddressBook {
         persons = new UniquePersonList();
     }
 
-    public AddressBook() {}
+    public PersonnelDatabase() {}
 
     /**
-     * Creates an AddressBook using the Persons in the {@code toBeCopied}
+     * Creates an PersonnelDatabase using the Persons in the {@code toBeCopied}
      */
-    public AddressBook(ReadOnlyAddressBook toBeCopied) {
+    public PersonnelDatabase(ReadOnlyPersonnelDatabase toBeCopied) {
         this();
         resetData(toBeCopied);
     }
@@ -52,15 +52,15 @@ public class AddressBook implements ReadOnlyAddressBook {
     }
 
     /**
-     * Resets the existing data of this {@code AddressBook} with {@code newData}.
+     * Resets the existing data of this {@code PersonnelDatabase} with {@code newData}.
      */
-    public void resetData(ReadOnlyAddressBook newData) {
+    public void resetData(ReadOnlyPersonnelDatabase newData) {
         requireNonNull(newData);
 
         setPersons(newData.getPersonList());
     }
 
-    // Sort the persons in the address book by name
+    // Sort the persons in the personnel database by name
     public void sort() {
         this.persons.sort();
     }
@@ -68,7 +68,7 @@ public class AddressBook implements ReadOnlyAddressBook {
     //// person-level operations
 
     /**
-     * Returns true if a person with the same identity as {@code person} exists in the address book.
+     * Returns true if a person with the same identity as {@code person} exists in the personnel database.
      */
     public boolean hasPerson(Person person) {
         requireNonNull(person);
@@ -76,8 +76,8 @@ public class AddressBook implements ReadOnlyAddressBook {
     }
 
     /**
-     * Adds a person to the address book.
-     * The person must not already exist in the address book.
+     * Adds a person to the personnel database.
+     * The person must not already exist in the personnel database.
      */
     public void addPerson(Person p) {
         persons.add(p);
@@ -86,8 +86,8 @@ public class AddressBook implements ReadOnlyAddressBook {
 
     /**
      * Replaces the given person {@code target} in the list with {@code editedPerson}.
-     * {@code target} must exist in the address book.
-     * The person identity of {@code editedPerson} must not be the same as another existing person in the address book.
+     * {@code target} must exist in the personnel database.
+     * The person identity of {@code editedPerson} must not be the same as another existing person in the personnel database.
      */
     public void setPerson(Person target, Person editedPerson) {
         requireNonNull(editedPerson);
@@ -97,8 +97,8 @@ public class AddressBook implements ReadOnlyAddressBook {
     }
 
     /**
-     * Removes {@code key} from this {@code AddressBook}.
-     * {@code key} must exist in the address book.
+     * Removes {@code key} from this {@code PersonnelDatabase}.
+     * {@code key} must exist in the personnel database.
      */
     public void removePerson(Person key) {
         persons.remove(key);
@@ -116,7 +116,7 @@ public class AddressBook implements ReadOnlyAddressBook {
     }
 
     /**
-     * Notifies listeners that the address book has been modified.
+     * Notifies listeners that the personnel database has been modified.
      */
     protected void indicateModified() {
         invalidationListenerManager.callListeners(this);
@@ -138,8 +138,8 @@ public class AddressBook implements ReadOnlyAddressBook {
     @Override
     public boolean equals(Object other) {
         return other == this // short circuit if same object
-                || (other instanceof AddressBook // instanceof handles nulls
-                && persons.equals(((AddressBook) other).persons));
+                || (other instanceof PersonnelDatabase // instanceof handles nulls
+                && persons.equals(((PersonnelDatabase) other).persons));
     }
 
     @Override
